+++ Core/World/Geometry/Sectors/Sector.cs
@@ -299,9 +299,16 @@ public sealed class Sector : SectorSoundSource, IFloorCeilingAnchor
             return;
 
         if (set)
+        {
             SectorEffect |= SectorEffect.Secret;
+            SectorEffect &= ~SectorEffect.FoundSecret;
+        }
         else
+        {
             SectorEffect &= ~SectorEffect.Secret;
+            SectorEffect |= SectorEffect.FoundSecret;
+        }
+
         DataChanges |= SectorDataTypes.SectorEffect;
     }